backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
Pillow>=10.0
numpy>=1.24
ImageHash>=4.3.1

# Optional — install if you want face clustering.
# Needs dlib which has native build deps (cmake, C++ toolchain).
# face_recognition>=1.3.0
